Feat: 경로 그리기 컴포넌트에서 훅으로 변경 (#266)
useDrawPolyline 훅 사용 방법

1.걷기 종료 버튼을 눌러서 데이터를 POST하는 컴포넌트의 리턴문에
<canvas ref={canvasRef} className={styles.canvasElement}/> 추가

2. scss에서 canvasElement 클래스에 대해 display: none

3. const { img, canvasRef } = useDrawPolyline(coordinates)

4. console.log(img) -> 이미지 url
