--- client/src/components/HistoryList/DrawPolyline.module.scss
@@ -1,10 +1,0 @@
-.canvasWrapper {
-  width: 64px;
-  height: 64px;
-  padding: 0px;
-  border: 1px solid red;
-
-  canvas {
-    border-radius: 50%;
-  }
-}
