--- client/src/components/HistoryList/dummyCoordinates.ts
@@ -1,40 +1,0 @@
-type CoordinateType = {
-  lat: number
-  lng: number
-}
-
-// export const coordinates: CoordinateType[] = [
-//   { lat: 37.579617, lng: 126.976433 }, // 광화문
-//   { lat: 37.579629, lng: 126.976842 },
-//   { lat: 37.579748, lng: 126.977237 },
-//   { lat: 37.579962, lng: 126.977581 },
-//   { lat: 37.580245, lng: 126.977852 },
-//   { lat: 37.58057, lng: 126.978034 },
-//   { lat: 37.58091, lng: 126.978114 },
-//   { lat: 37.581238, lng: 126.978085 },
-//   { lat: 37.581527, lng: 126.977949 },
-//   { lat: 37.58176, lng: 126.977714 },
-//   { lat: 37.58192, lng: 126.9774 },
-//   { lat: 37.582, lng: 126.977044 },
-//   { lat: 37.581986, lng: 126.976678 },
-//   { lat: 37.58188, lng: 126.976345 },
-//   { lat: 37.581693, lng: 126.976068 },
-//   { lat: 37.58144, lng: 126.975869 },
-//   { lat: 37.581135, lng: 126.975762 },
-//   { lat: 37.580798, lng: 126.975754 },
-//   { lat: 37.580453, lng: 126.975846 },
-//   { lat: 37.580121, lng: 126.976027 },
-//   { lat: 37.579819, lng: 126.976287 },
-//   { lat: 37.579617, lng: 126.976433 }, // 광화문 (시작점과 끝점 동일)
-// ]
-
-export const coordinates: CoordinateType[] = [
-  { lat: 37.5558, lng: 126.9725 }, // 서울역
-  { lat: 37.5789, lng: 126.9436 },
-  { lat: 37.582, lng: 126.9518 },
-  { lat: 37.5842, lng: 126.9591 },
-  { lat: 37.5906, lng: 126.9589 },
-  { lat: 37.5939, lng: 126.9534 },
-  { lat: 37.5951, lng: 126.9514 },
-  { lat: 37.5989, lng: 126.9496 }, // 서울시립대학교
-]
