+++ client/src/hooks/dummyCoordinates.ts
@@ -1,0 +1,15 @@
+type CoordinateType = {
+  lat: number
+  lng: number
+}
+
+export const coordinates: CoordinateType[] = [
+  { lat: 37.5558, lng: 126.9725 }, // 서울역
+  { lat: 37.5789, lng: 126.9436 },
+  { lat: 37.582, lng: 126.9518 },
+  { lat: 37.5842, lng: 126.9591 },
+  { lat: 37.5906, lng: 126.9589 },
+  { lat: 37.5939, lng: 126.9534 },
+  { lat: 37.5951, lng: 126.9514 },
+  { lat: 37.5989, lng: 126.9496 }, // 서울시립대학교
+]
